+++ src/move_gen/pseudo_legal_moves.rs
@@ -54,7 +54,7 @@ impl PseudoLegalMoveGenerator {
             let n = precomp.num_sqrs_to_edge[attack_sqr.index()][dir];
             let direction_offset = precomp.direction_offsets[dir];
             for i in 0..n {
-                let sqr_idx = attack_sqr.square() as i32 + direction_offset * (i + 1);
+                let sqr_idx = attack_sqr.square() + direction_offset * (i + 1);
                 let piece = board.square[sqr_idx as usize];
                 if piece != Piece::NULL {
                     if piece.is_color(friendly_color) { break; }
@@ -77,7 +77,7 @@ impl PseudoLegalMoveGenerator {
         };
         for i in 0..2 {
             if precomp.num_sqrs_to_edge[attack_sqr.index()][precomp.pawn_attack_dirs[friendly_color_idx][i].index()] > 0 {
-                let s = attack_sqr.square() as i32 + precomp.direction_offsets[precomp.pawn_attack_dirs[friendly_color_idx][i].index()];
+                let s = attack_sqr.square() + precomp.direction_offsets[precomp.pawn_attack_dirs[friendly_color_idx][i].index()];
                 let piece = board.square[s as usize];
                 if piece.value() == (Piece::PAWN | attacker_color) { return true; }
             }
@@ -145,14 +145,14 @@ impl PseudoLegalMoveGenerator {
         for dir_idx in start_dir_idx..end_dir_idx {
             let cur_dir_offset = precomp.direction_offsets[dir_idx];
             for n in 0..precomp.num_sqrs_to_edge[start_sqr.index()][dir_idx] {
-                let target_sqr = start_sqr.square() as i32 + cur_dir_offset * (n + 1);
+                let target_sqr = start_sqr.square() + cur_dir_offset * (n + 1);
                 let target_sqr_piece = board.square[target_sqr as usize];
 
                 if target_sqr_piece.is_color(self.friendly_color) { break; }
 
                 let is_capture = target_sqr_piece != Piece::NULL;
                 if self.gen_quiets || is_capture {
-                    self.moves.push(Move::from_start_end(start_sqr.square(), target_sqr as u8));
+                    self.moves.push(Move::from_start_end(start_sqr.square(), target_sqr));
                 }
                 if is_capture { break; }
             }
@@ -179,8 +179,8 @@ impl PseudoLegalMoveGenerator {
         let start_rank = if self.white_to_move { 1 } else { 6 };
         let final_rank_before_prom = if self.white_to_move { 6 } else { 1 };
 
-        let en_passant_file = ((board.current_game_state >> 4) & 15) as i32 - 1;
-        let mut en_passant_sqr = -1;
+        let en_passant_file = ((board.current_game_state >> 4) & 15) as i8 - 1;
+        let mut en_passant_sqr = -1i8;
         if en_passant_file != -1 {
             en_passant_sqr = 8 * (if self.white_to_move { 5 } else { 2 }) + en_passant_file;
         }
@@ -191,17 +191,17 @@ impl PseudoLegalMoveGenerator {
             let one_step_from_prom = rank == final_rank_before_prom;
 
             if self.gen_quiets {
-                let square_one_forward = start_sqr.square() as i32 + pawn_offset;
+                let square_one_forward = start_sqr.square() + pawn_offset;
                 if board.square[square_one_forward as usize] == Piece::NULL {
                     if one_step_from_prom {
-                        self.make_promotion_moves(start_sqr, Coord::from_idx(square_one_forward as u8));
+                        self.make_promotion_moves(start_sqr, Coord::from_idx(square_one_forward));
                     } else {
-                        self.moves.push(Move::from_start_end(start_sqr.square(), square_one_forward as u8));
+                        self.moves.push(Move::from_start_end(start_sqr.square(), square_one_forward));
                     }
                     if rank == start_rank {
                         let square_two_forward = square_one_forward + pawn_offset;
                         if board.square[square_two_forward as usize] == Piece::NULL {
-                            self.moves.push(Move::from_start_end_flagged(start_sqr.square(), square_two_forward as u8, Move::PAWN_TWO_FORWARD));
+                            self.moves.push(Move::from_start_end_flagged(start_sqr.square(), square_two_forward, Move::PAWN_TWO_FORWARD));
                         }
                     }
                 }
@@ -210,19 +210,19 @@ impl PseudoLegalMoveGenerator {
             for j in 0..2 {
                 if precomp.num_sqrs_to_edge[start_sqr.index()][precomp.pawn_attack_dirs[self.friendly_color_idx][j].index()] > 0 {
                     let pawn_capture_dir = precomp.direction_offsets[precomp.pawn_attack_dirs[self.friendly_color_idx][j].index()];
-                    let target_sqr = start_sqr.square() as i32 + pawn_capture_dir;
+                    let target_sqr = start_sqr.square() + pawn_capture_dir;
                     let target_piece = board.square[target_sqr as usize];
 
                     if target_piece.is_color(self.opponent_color) {
                         if one_step_from_prom {
-                            self.make_promotion_moves(start_sqr, Coord::from_idx(target_sqr as u8));
+                            self.make_promotion_moves(start_sqr, Coord::from_idx(target_sqr));
                         } else {
-                            self.moves.push(Move::from_start_end(start_sqr.square(), target_sqr as u8));
+                            self.moves.push(Move::from_start_end(start_sqr.square(), target_sqr));
                         }
                     }
                     if target_sqr == en_passant_sqr {
                         let ep_captured_pawn_sqr = target_sqr + if self.white_to_move { -8 } else { 8 };
-                        self.moves.push(Move::from_start_end_flagged(start_sqr.square(), target_sqr as u8, Move::EN_PASSANT_CAPTURE));
+                        self.moves.push(Move::from_start_end_flagged(start_sqr.square(), target_sqr, Move::EN_PASSANT_CAPTURE));
                     }
                 }
             }
